Convenience Script to download training models
Previously this was done in the Dockerfile, which was annoying as each
time it is built it downloads the same file over and over again. Now
this has been externalized into a separate script so users can now just
download their own training models on their own or just use this script
to do so.