Removed bitrate conversion from ffmpeg
Why is this required? I saw no indication on the deepspeech docs that a
particular bitrate was favored. To reduce chance of artifacts I believe
this should just be kept as whatever birate the source used. The
frequency resampling with a single mono channel is all that appears
to be required for the model. Correct me if I am wrong, but I couldn't
find any documentation that this bitrate was favored.