Remove SRT/VTT File if it exists
Since this likely indicates an attempt to re-run the program potentially
with different settings or with different models, it appears safe to
just delete it. As I was testing the code, I had to re-run multiple
times and hence it was convenient to just have the thing delete it for
me. A potential alternative if this is unacceptable is to rename the
file with a randomly generated string to ensure it isn't deleted while
still allowing for a seamless re-running of the program.

Also, there was a redundant line at the top of the commit that needed to
be removed.