Fix extra colon character when time ends in '00'
When subtitles the millisecond marker ends in ".00", the except block does not
get triggered as appears to have been intended.

This bug causes the subtitles contain an unexpected colon character instead of
the final digit, and that can cause parsing issues when the SRT file is read by
external programs.

Fixes #31