Split run-on sentences into multiple subtitles
Add --split-duration argument to split up the inferred subtitles based on a
configured number of seconds (default: 5 seconds).

Recall that Autosub splits up the input audio based silent sections. For some
input there may be tens of seconds (or even minutes) of audio without any
silence, which leads Autosub to produce a very long line. This commit adds a
command-line parameter and a sensible default so that the subtitles sentences
are naturally split up in a manner that can be read at standard reading speed.

Fixes #21