Drops --vtt, create SRT/VTT unless --format used
Replaces the --vtt argument that is used to switch between creating an SRT file
or a VTT file, to instead always create both formats by default and allow the
user to specify a subset of formats using the --format argument introduced in
this commit.

As discussed in #33, it is the inference part that takes the bulk of the
execution time and the output subtitle file format doesn't take much time and
uses very little disk space then rerunning the entire inference again just to
get a VTT file is inefficient, so it makes more sense just to create the VTT
and SRT by default.

Fixes #33