Handle input filenames containing special chars
Improves the handling of filenames which contain special characters by not
processing the ffmpeg command using a shell, and passing the arguments in using
a Python list.

Note: The use of the old 'subprocess.call()' API here is unchanged. A future
commit may switch to the more modern subprocess.run().